# Apply micro-optimisation for Nearest API